TH 🍋
This is a tricky one sorta since lemon is [เลมอน](http://learn-thai.portalsbay.com/index.php?page=learnthai&id=15335&c=32&englishlanguage=lemon ), and while มะนาว is lime, it's often used as a stand-in for lemon as the distinction between the two isn't really made often. I think I could compare it to me saying basil is basil, but Thais always make the distinction between [Thai basil and holy basil](https://www.spiceography.com/thai-basil-vs-holy-basil/) (which doesn't include the sweet basil); I used to use Thai basil when cooking in the US since the Asians store sold it for cheaper than chain basils haha.  ชามะนาว is a lime tea, but is people often translate it to lemon tea. I'm just guessing since lemon tea is much more common in the West that's what the translations bots say and it's all over the signs, but it's definitely lime. I usually try to ask if it's lemon or lime if I order off a menu, and people seem puzzled that I'd even ask. That said, I much prefer lime tea. /story

@turboMaCk